Fix display of right paren placeholders on calc screen

+++ js/app/viewModel.js
@@ -267,7 +267,7 @@ define(['view', 'expression', 'inputFilter', 'math'], function(view, expression,
 				//configure math.js lib to prevent round-off errors.
 				math.config({
 					number: 'BigNumber',
-					precision: 16 // display up to 15 numbers
+					precision: 15 // number of digits to display
 				});	
 				
 				function addListeners() {
